+++ src/BDSDetectorConstruction.cc
@@ -58,6 +58,7 @@ along with BDSIM.  If not, see <http://www.gnu.org/licenses/>.
 #include "BDSScorerInfo.hh"
 #include "BDSScorerMeshInfo.hh"
 #include "BDSScoringMeshBox.hh"
+#include "BDSScoringMeshCylinder.hh"
 #include "BDSSDEnergyDeposition.hh"
 #include "BDSSDManager.hh"
 #include "BDSSDType.hh"
@@ -1238,18 +1239,34 @@ void BDSDetectorConstruction::ConstructScoringMeshes()
       // TBC - could be any beam line in future - just w.r.t. main beam line just now
       const BDSBeamline* mbl = BDSAcceleratorModel::Instance()->BeamlineMain();
       G4Transform3D placement = CreatePlacementTransform(mesh, mbl);
-      
+
+      BDSScoringMeshBox* scorerBox = nullptr;
+      BDSScoringMeshCylinder* scorerCylindrical = nullptr;
+      const BDSHistBinMapper* mapper = nullptr;
+
+      G4String geometryType = BDS::LowerCase(G4String(mesh.geometryType));
+
+      if (geometryType == "box") {
       // create a scoring box
-      BDSScoringMeshBox* scorerBox = new BDSScoringMeshBox(meshName, meshRecipe, placement);
-      const BDSHistBinMapper* mapper = scorerBox->Mapper();
-      
+        scorerBox = new BDSScoringMeshBox(meshName, meshRecipe, placement);
+        mapper = scorerBox->Mapper();}
+
+      else if (geometryType == "cylindrical") {
+      // create a scoring cylinder
+        scorerCylindrical = new BDSScoringMeshCylinder(meshName, meshRecipe, placement);
+        mapper = scorerCylindrical->Mapper();}
+
+      else {
+          throw BDSException(__METHOD_NAME__, "mesh geometry type \"" + geometryType + "\" is not correct. The possible options are \"box\" and \"cylindrical\"");}
+
       // add the scorer(s) to the scoring mesh
       std::vector<G4String> meshPrimitiveScorerNames; // final vector of unique mesh + ps names
       std::vector<G4double> meshPrimitiveScorerUnits;
       std::vector<G4String> scorerNames;
       std::stringstream sqss(mesh.scoreQuantity);
       G4String word;
-      while (sqss >> word) // split by white space - process word at a time
+
+        while (sqss >> word) // split by white space - process word at a time
 	{
 	  auto search = scorerRecipes.find(word);
 	  if (search == scorerRecipes.end())
@@ -1264,13 +1281,21 @@ void BDSDetectorConstruction::ConstructScoringMeshes()
 	  G4String uniqueName = meshName + "/" + ps->GetName();
 	  meshPrimitiveScorerNames.push_back(uniqueName);
 	  meshPrimitiveScorerUnits.push_back(psUnit);
-	  scorerBox->SetPrimitiveScorer(ps); // sets the current ps but appends to list of multiple
+
+      if (geometryType == "box"){
+          scorerBox->SetPrimitiveScorer(ps);} // sets the current ps but appends to list of multiple
+      else if (geometryType == "cylindrical"){
+          scorerCylindrical->SetPrimitiveScorer(ps);}// sets the current ps but appends to list of multiple
+
 	  BDSScorerHistogramDef outputHistogram(meshRecipe, uniqueName, ps->GetName(), psUnit, *mapper);
 	  BDSAcceleratorModel::Instance()->RegisterScorerHistogramDefinition(outputHistogram);
 	  BDSAcceleratorModel::Instance()->RegisterScorerPlacement(meshName, placement);
 	}
 
-      scManager->RegisterScoringMesh(scorerBox);
+      if (geometryType == "box"){
+          scManager->RegisterScoringMesh(scorerBox);} // sets the current ps but appends to list of multiple
+      else if (geometryType == "cylindrical"){
+          scManager->RegisterScoringMesh(scorerCylindrical);}// sets the current ps but appends to list of multiple
 
       // register it with the sd manager as this is where we get all collection IDs from
       // in the end of event action. This must come from the mesh as it creates the